split monitor widget in different files to improve readability and testability
